requirements.txt
futures>=3.2.0; python_version < "3.0"
google-api-python-client>=1.7.3
google-auth>=1.5.0
google-auth-httplib2>=0.0.3
iso8601>=0.1.12
six>=1.11.0
